backend/pyproject.toml
[tool.poetry]
name = "snapdocs-backend"
version = "0.1.0"
description = "SnapDocs backend (FastAPI + MongoDB)"
authors = ["You <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.6" }
python-dotenv = "^1.0.1"
motor = "^3.5.1"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
PyJWT = "^2.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
